Minor updates to import/export scripts

+++ utils/importData.py
@@ -10,19 +10,25 @@ import os, sys, shutil, glob, argparse
 from subprocess import call
 
 parser = argparse.ArgumentParser(description="Import DCMetroMetrics data into MongoDB")
+parser.add_argument('--db', default="MetroEscalators", help='The database to use.')
+parser.add_argument('--zip', default="./DCMetroMetricsData", help='The zip file to use')
 
 c = MongoClient()
-db = c.MetroEscalators
+
 
 def run():
     args = parser.parse_args()
 
-    ZIP = 'DCMetroMetricsData.zip'
+    ZIP = args.zip
+
+    if not os.path.exists(ZIP):
+        raise RuntimeError("Could not find zip file: %s"%ZIP)
+
     UNZIPPED_DIR = os.path.abspath('./DCMetroMetricsData')
 
     # Download the latest data zip file
-    cmd = "curl http://www.dcmetrometrics.com/data/{ZIP} > {ZIP}".format(ZIP=ZIP)
-    call(cmd, shell=True)
+    # cmd = "curl http://www.dcmetrometrics.com/data/{ZIP} > {ZIP}".format(ZIP=ZIP)
+    # call(cmd, shell=True)
 
     if os.path.exists(UNZIPPED_DIR):
         shutil.rmtree(UNZIPPED_DIR)
@@ -31,6 +37,7 @@ def run():
     cmd = "unzip %s"%ZIP
     call(cmd, shell=True)
 
+    db = c[args.db]
 
     jsons = glob.glob('%s/*.json'%UNZIPPED_DIR)
     print 'Json Directory: %s'%UNZIPPED_DIR
@@ -47,8 +54,8 @@ def run():
         print 'Dropping collection %s'%collection
         db[collection].drop()
         print 'Importing json %s into collection %s'%(j, collection)
-        cmd = 'mongoimport -d MetroEscalators -c {collection} {fname}'
-        cmd = cmd.format(collection=collection, fname=j)
+        cmd = 'mongoimport -d {db} -c {collection} {fname}'
+        cmd = cmd.format(db = args.db, collection=collection, fname=j)
         call(cmd, shell=True)
 
 if __name__ == '__main__':
